requirements.txt
Flask
requests
tweepy
google-api-python-client
oauthlib
requests-oauthlib
PyYAML
python-dotenv
flask-cors
Flask-SQLAlchemy
Flask-Migrate
Flask-RESTful
Flask-Login
Flask-Mail
Flask-WTF
textblob
nltk
Flask-PyMongo